requirements.txt
python-dotenv
beautifulsoup4
httpx
firebase_admin